Refactor intcode emulator, fix all usages

+++ src/lib/include/io/io.hpp
@@ -5,6 +5,6 @@
 
 namespace io {
 std::vector<int> ReadIntegers(const std::string& path);
-std::vector<int> ReadInstructions(const std::string& path);
+std::vector<int64_t> ReadInstructions(const std::string& path);
 }
